Copy value of key rather than saving pointer
This ensures that when the data at the provided key's address is changed,
the key in the hashmap does not change.

This would break the integrity of the hashmap.

Because c99 does not support the strdup function, it is declared
at the top of the file to ensure correct behaviour when compiled
with the Makefile of the assignment.